Move BufferedFileReader, FileReader, and StringReader sources into the clp namespace.
