+++ components/core/src/clp/Grep.cpp
@@ -7,9 +7,9 @@
 
 #include "../ir/parsing.hpp"
 #include "../ir/types.hpp"
-#include "../StringReader.hpp"
 #include "EncodedVariableInterpreter.hpp"
 #include "LogSurgeonReader.hpp"
+#include "StringReader.hpp"
 #include "Utils.hpp"
 
 using clp::streaming_archive::reader::Archive;
