+++ components/core/src/clp/Utils.hpp
@@ -11,7 +11,7 @@
 
 #include "../Defs.h"
 #include "../ErrorCode.hpp"
-#include "../FileReader.hpp"
+#include "FileReader.hpp"
 #include "ParsedMessage.hpp"
 
 namespace clp {
