+++ components/core/src/clp/VariableDictionaryEntry.hpp
@@ -3,8 +3,8 @@
 
 #include "../Defs.h"
 #include "../ErrorCode.hpp"
-#include "../FileReader.hpp"
 #include "DictionaryEntry.hpp"
+#include "FileReader.hpp"
 #include "streaming_compression/zstd/Compressor.hpp"
 #include "streaming_compression/zstd/Decompressor.hpp"
