+++ components/core/src/clp/streaming_compression/zstd/Decompressor.hpp
@@ -7,8 +7,8 @@
 #include <boost/iostreams/device/mapped_file.hpp>
 #include <zstd.h>
 
-#include "../../../FileReader.hpp"
 #include "../../../TraceableException.hpp"
+#include "../../FileReader.hpp"
 #include "../Decompressor.hpp"
 
 namespace clp::streaming_compression::zstd {
